cainescrossfire: 64-bit squared distances, a harness launch, and the specials' own logs
The hornet ring's contact test computed dx*dx + dz*dz in 32 bits. A squared
world distance overflows that on a big map, and a wrapped NEGATIVE compared as
"inside the ring" - the ring was damaging and knocking cars on the far side of
the level (14 hits in a run, 13 of them nonsense; 1 real one left). Every other
squared-distance test in the module had the same shape and the same failure, but
as a MINIMUM search: a wrapped distance reads as the NEAREST car, so Corvo's
bolt, the missiles' homing pick, the AI's target/obstacle/separation tests and
the raycast threat scan could all pick a car across the map. All of them, and
the three integer sqrt helpers they feed, are 64-bit now.

Also here, because it is what proved the above:

- the CC select flow gets a harness launch: with CC_FORCE_ARENA and CC_FORCE_CAR
  set the match starts from the frame handler instead of on a menu confirm,
  since a padless run cannot drive the menus at all. A pad still gets the menus.
- obelisk/avalanche/hornet say what they did in the log (salvo fired and done,
  crush latched and pushed off, ring contact), which is how a headless run sees
  a special that has no input of its own.
- the debug driver knows special_obelisk (only the bare "obelisk" name was
  added with the car).

+++ JERICHO/MODS/cainescrossfire/weapons/raycast/raycast.c
@@ -230,7 +230,7 @@ int cd2RaycastThreat(const CAR_DATA* car, VECTOR* pos, VECTOR* vel)
 	for (i = 0; i < CD2_MAX_RAYCAST; i++)
 	{
 		CD2_RAYCAST* r = &gRcast[i];
-		int dx, dy, dz, dist2, closing;
+		long long dx, dy, dz, dist2, closing;  /* 64-bit: the squared sum wraps otherwise */
 
 		if (!r->active || r->owner == car)
 			continue;
